+++ src/theme/dashboard-navbar.ts
@@ -5,7 +5,7 @@
  */
 export default {
   slots: {
-    root: 'h-(--b24ui-header-height) shrink-0 flex items-center justify-between border-b border-(--ui-color-divider-default) px-4 sm:px-6 gap-1.5',
+    root: 'h-(--b24ui-header-height) shrink-0 flex items-center justify-between border-b border-(--ui-color-divider-default) px-4 gap-1.5',
     left: 'flex items-center gap-1.5 min-w-0',
     icon: 'shrink-0 size-5 self-center me-1.5',
     title: 'flex items-center gap-1.5 font-(--ui-font-weight-semi-bold) text-(--b24ui-typography-label-color) truncate',