+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * fix(Page): make slot presence reactive for variant computation
 * fix(useResizable): use function declaration to prevent false auto-import
 * fix(ContentToc): add relative positioning to content slot
+* fix(components): improve arrow styling with `stroke-default` and `fill-bg`
 * docs(navigation-menu): improve examples
 * docs(input): add phone number example